Q2/data-extraction-project/requirements.txt
# requirements.txt for Apache Airflow

# Apache Airflow
apache-airflow==2.7.0

# Additional packages for your DAGs and plugins
pandas
numpy
requests
usaddress
praw
usaddress==0.5.10
scikit-learn==1.3.0
sqlalchemy
mysql-connector-python
validators
nltk
google-api-python-client